sphere intersection

+++ minirt/headers/utils.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   utils.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: aloubry <aloubry@student.42.fr>            +#+  +:+       +#+        */
+/*   By: lpittet <lpittet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/14 16:48:25 by aloubry           #+#    #+#             */
-/*   Updated: 2025/02/24 23:34:27 by aloubry          ###   ########.fr       */
+/*   Updated: 2025/02/26 15:17:05 by lpittet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ t_vector3	v3_normalize(t_vector3 v);
 
 // vectors2.c
 
-double		v3_get_magnitude(t_vector3 v);
+float		v3_get_magnitude(t_vector3 v);
 float		v3_dot(t_vector3 a, t_vector3 b);
 t_vector3	v3_up(void);
